Move formElementRank and formElementSpaceDim into formutils.py
